refactor: reorganize parameter parsing and http error handling
