+++ aio/aio-proxy/aio_proxy/parsers/tranche_effectif.py
@@ -1,0 +1,28 @@
+from typing import Optional
+
+from aio_proxy.labels.helpers import tranches_effectifs
+
+
+def validate_tranche_effectif_salarie_entreprise(
+    tranche_effectif_salarie_entreprise_clean: str,
+) -> Optional[str]:
+    """Check the validity of tranche_effectif_salarie_entreprise.
+
+    Args:
+        tranche_effectif_salarie_entreprise_clean(str, optional):
+         tranche_effectif_salarie_entreprise extracted and cleaned.
+
+    Returns:
+        None if tranche_effectif_salarie_entreprise_clean is None.
+        tranche_effectif_salarie_entreprise_clean if valid.
+
+    Raises:
+        ValueError: if tranche_effectif_salarie_entreprise_clean not valid.
+    """
+    if tranche_effectif_salarie_entreprise_clean is None:
+        return None
+    if len(tranche_effectif_salarie_entreprise_clean) != 2:
+        raise ValueError("Tranche salariés doit contenir 2 caractères.")
+    if tranche_effectif_salarie_entreprise_clean not in tranches_effectifs:
+        raise ValueError("Tranche salariés non valide.")
+    return tranche_effectif_salarie_entreprise_clean
