+++ aio/aio-proxy/aio_proxy/routes.py
@@ -2,10 +2,7 @@ import json
 import os
 
 import sentry_sdk
-from aio_proxy.parameters import (
-    extract_geo_parameters,
-    extract_text_parameters,
-)
+from aio_proxy.parameters import extract_geo_parameters, extract_text_parameters
 from aio_proxy.response import api_response
 from aio_proxy.search.search_functions import search_geo, search_text
 from aiohttp import web
